Move test_replace to directories
