+++ python/cudf/cudf/tests/series/methods/test_replace.py
@@ -1,5 +1,6 @@
 # Copyright (c) 2025, NVIDIA CORPORATION.
 
+import re
 
 import numpy as np
 import pandas as pd
@@ -7,11 +8,14 @@ import pytest
 
 import cudf
 from cudf.core._compat import (
+    PANDAS_CURRENT_SUPPORTED_VERSION,
     PANDAS_GE_220,
     PANDAS_GT_214,
+    PANDAS_VERSION,
 )
 from cudf.testing import assert_eq
 from cudf.testing._utils import (
+    assert_exceptions_equal,
     expect_warning_if,
 )
 
@@ -278,3 +282,246 @@ def test_numeric_series_replace_dtype(
     expect = psr.replace([2, 3], [replacement, replacement])
     got = sr.replace([2, 3], [replacement, replacement])
     assert_eq(expect, got)
+
+
+@pytest.mark.parametrize(
+    "pframe, replace_args",
+    [
+        (
+            pd.Series([5, 1, 2, 3, 4]),
+            {"to_replace": 5, "value": 0, "inplace": True},
+        ),
+        (
+            pd.Series([5, 1, 2, 3, 4]),
+            {"to_replace": {5: 0, 3: -5}, "inplace": True},
+        ),
+        (pd.Series([5, 1, 2, 3, 4]), {}),
+        pytest.param(
+            pd.Series(["one", "two", "three"], dtype="category"),
+            {"to_replace": "one", "value": "two", "inplace": True},
+            marks=pytest.mark.xfail(
+                condition=PANDAS_VERSION >= PANDAS_CURRENT_SUPPORTED_VERSION,
+                reason="https://github.com/pandas-dev/pandas/issues/43232"
+                "https://github.com/pandas-dev/pandas/issues/53358",
+            ),
+        ),
+        (
+            pd.DataFrame({"A": [0, 1, 2, 3, 4], "B": [5, 6, 7, 8, 9]}),
+            {"to_replace": 5, "value": 0, "inplace": True},
+        ),
+        (
+            pd.Series([1, 2, 3, 45]),
+            {
+                "to_replace": np.array([]).astype(int),
+                "value": 77,
+                "inplace": True,
+            },
+        ),
+        (
+            pd.Series([1, 2, 3, 45]),
+            {
+                "to_replace": np.array([]).astype(int),
+                "value": 77,
+                "inplace": False,
+            },
+        ),
+        (
+            pd.DataFrame({"a": [1, 2, 3, 4, 5, 666]}),
+            {"to_replace": {"a": 2}, "value": {"a": -33}, "inplace": True},
+        ),
+        (
+            pd.DataFrame({"a": [1, 2, 3, 4, 5, 666]}),
+            {
+                "to_replace": {"a": [2, 5]},
+                "value": {"a": [9, 10]},
+                "inplace": True,
+            },
+        ),
+        (
+            pd.DataFrame({"a": [1, 2, 3, 4, 5, 666]}),
+            {"to_replace": [], "value": [], "inplace": True},
+        ),
+    ],
+)
+@pytest.mark.skipif(
+    PANDAS_VERSION < PANDAS_CURRENT_SUPPORTED_VERSION,
+    reason="Warning not given on older versions of pandas",
+)
+def test_replace_inplace(pframe, replace_args):
+    gpu_frame = cudf.from_pandas(pframe)
+    pandas_frame = pframe.copy()
+
+    gpu_copy = gpu_frame.copy()
+    cpu_copy = pandas_frame.copy()
+
+    assert_eq(gpu_frame, pandas_frame)
+    assert_eq(gpu_copy, cpu_copy)
+    with expect_warning_if(len(replace_args) == 0):
+        gpu_frame.replace(**replace_args)
+    with expect_warning_if(len(replace_args) == 0):
+        pandas_frame.replace(**replace_args)
+    assert_eq(gpu_frame, pandas_frame)
+    assert_eq(gpu_copy, cpu_copy)
+
+
+def test_series_replace_errors():
+    gsr = cudf.Series([1, 2, None, 3, None])
+    psr = gsr.to_pandas()
+
+    with pytest.raises(
+        TypeError,
+        match=re.escape(
+            "to_replace and value should be of same types,"
+            "got to_replace dtype: int64 and "
+            "value dtype: object"
+        ),
+    ):
+        gsr.replace(1, "a")
+
+    gsr = cudf.Series(["a", "b", "c"])
+    with pytest.raises(
+        TypeError,
+        match=re.escape(
+            "to_replace and value should be of same types,"
+            "got to_replace dtype: int64 and "
+            "value dtype: object"
+        ),
+    ):
+        gsr.replace([1, 2], ["a", "b"])
+
+    assert_exceptions_equal(
+        psr.replace,
+        gsr.replace,
+        ([{"a": 1}, 1],),
+        ([{"a": 1}, 1],),
+    )
+
+    assert_exceptions_equal(
+        lfunc=psr.replace,
+        rfunc=gsr.replace,
+        lfunc_args_and_kwargs=([[1, 2], [1]],),
+        rfunc_args_and_kwargs=([[1, 2], [1]],),
+    )
+
+    assert_exceptions_equal(
+        lfunc=psr.replace,
+        rfunc=gsr.replace,
+        lfunc_args_and_kwargs=([object(), [1]],),
+        rfunc_args_and_kwargs=([object(), [1]],),
+    )
+
+    assert_exceptions_equal(
+        lfunc=psr.replace,
+        rfunc=gsr.replace,
+        lfunc_args_and_kwargs=([{"a": 1}, object()],),
+        rfunc_args_and_kwargs=([{"a": 1}, object()],),
+    )
+
+
+@pytest.mark.parametrize(
+    "gsr,old,new,expected",
+    [
+        (
+            lambda: cudf.Series(["a", "b", "c", None]),
+            None,
+            "a",
+            lambda: cudf.Series(["a", "b", "c", "a"]),
+        ),
+        (
+            lambda: cudf.Series(["a", "b", "c", None]),
+            [None, "a", "a"],
+            ["c", "b", "d"],
+            lambda: cudf.Series(["d", "b", "c", "c"]),
+        ),
+        (
+            lambda: cudf.Series(["a", "b", "c", None]),
+            [None, "a"],
+            ["b", None],
+            lambda: cudf.Series([None, "b", "c", "b"]),
+        ),
+        (
+            lambda: cudf.Series(["a", "b", "c", None]),
+            [None, None],
+            [None, None],
+            lambda: cudf.Series(["a", "b", "c", None]),
+        ),
+        (
+            lambda: cudf.Series([1, 2, None, 3]),
+            None,
+            10,
+            lambda: cudf.Series([1, 2, 10, 3]),
+        ),
+        (
+            lambda: cudf.Series([1, 2, None, 3]),
+            [None, 1, 1],
+            [3, 2, 4],
+            lambda: cudf.Series([4, 2, 3, 3]),
+        ),
+        (
+            lambda: cudf.Series([1, 2, None, 3]),
+            [None, 1],
+            [2, None],
+            lambda: cudf.Series([None, 2, 2, 3]),
+        ),
+        (
+            lambda: cudf.Series(["a", "q", "t", None], dtype="category"),
+            None,
+            "z",
+            lambda: cudf.Series(["a", "q", "t", "z"], dtype="category"),
+        ),
+        (
+            lambda: cudf.Series(["a", "q", "t", None], dtype="category"),
+            [None, "a", "q"],
+            ["z", None, None],
+            lambda: cudf.Series([None, None, "t", "z"], dtype="category"),
+        ),
+        (
+            lambda: cudf.Series(["a", None, "t", None], dtype="category"),
+            [None, "t"],
+            ["p", None],
+            lambda: cudf.Series(["a", "p", None, "p"], dtype="category"),
+        ),
+    ],
+)
+def test_replace_nulls(gsr, old, new, expected):
+    gsr = gsr()
+    with expect_warning_if(isinstance(gsr.dtype, cudf.CategoricalDtype)):
+        actual = gsr.replace(old, new)
+    assert_eq(
+        expected().sort_values().reset_index(drop=True),
+        actual.sort_values().reset_index(drop=True),
+    )
+
+
+def test_replace_with_index_objects():
+    result = cudf.Series([1, 2]).replace(cudf.Index([1]), cudf.Index([2]))
+    expected = pd.Series([1, 2]).replace(pd.Index([1]), pd.Index([2]))
+    assert_eq(result, expected)
+
+
+def test_replace_datetime_series():
+    pd_series = pd.Series(pd.date_range("20210101", periods=5))
+    pd_result = pd_series.replace(
+        pd.Timestamp("2021-01-02"), pd.Timestamp("2021-01-10")
+    )
+
+    cudf_series = cudf.Series(pd.date_range("20210101", periods=5))
+    cudf_result = cudf_series.replace(
+        pd.Timestamp("2021-01-02"), pd.Timestamp("2021-01-10")
+    )
+
+    assert_eq(pd_result, cudf_result)
+
+
+def test_replace_timedelta_series():
+    pd_series = pd.Series(pd.timedelta_range("1 days", periods=5))
+    pd_result = pd_series.replace(
+        pd.Timedelta("2 days"), pd.Timedelta("10 days")
+    )
+
+    cudf_series = cudf.Series(pd.timedelta_range("1 days", periods=5))
+    cudf_result = cudf_series.replace(
+        pd.Timedelta("2 days"), pd.Timedelta("10 days")
+    )
+
+    assert_eq(pd_result, cudf_result)
